requirements.txt
streamlit>=1.31.0
google-genai>=1.0.0
chromadb>=0.5.0
pypdf>=4.0.0
sentence-transformers>=3.0.0
scikit-learn>=1.4.0
networkx>=3.2
pyvis>=0.3.2
plotly>=5.18.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
